Utilities

RTI.meshgrid — Function

Same with meshgrid in Matlab

X,Y=meshgrid(x,y) returns 2-D grid coordinates based on the coordinates contained in vectors x and y. X is a matrix where each row is a copy of x, and Y is a matrix where each column is a copy of y. The grid represented by the coordinates X and Y has length(y) rows and length(x) columns.

X,Y,Z=meshgrid(x,y,z) returns 3-D grid coordinates defined by the vectors x, y, and z. The grid represented by X, Y, and Z has size length(y)-by-length(x)-by-length(z).

RTI.write2mat — Function

write a mat file.

Example: write2mat(path,var).

RTI.readmat — Function

read a mat file.

Example: readmat(path,var).

RTI.rickerWave — Function

ricker wavelet.

Example:rickerWave(frequency,dt,nt,magnitude).

RTI.eikonal.acoustic_eikonal_forward — Function

Forward eikonal equation solver.

RTI.eikonal.acoustic_eikonal_adjoint — Function

Adjoint eikonal equation solver.